requirements.txt
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
jinja2>=3.1.6
numpy>=1.21.0
pandas>=1.3.0
psycopg2-binary>=2.9.10
pydantic>=2.0.0
werkzeug>=3.1.3
cryptography>=3.4.8
email-validator>=1.1.3
python-multipart>=0.0.5
scikit-learn>=1.0.0
scipy>=1.7.0
python-dotenv>=1.0.0
openpyxl
xlrd
reportlab
python-docx
jinja2
